requirements.txt
python-telegram-bot==20.6
apscheduler==3.10.4
tzdata
aiohttp
